+++ src/player/main/sp/it/pl/gui/itemnode/textfield/FontTextField.kt
@@ -2,6 +2,7 @@ package sp.it.pl.gui.itemnode.textfield
 
 import javafx.scene.text.Font
 import sp.it.pl.gui.objects.picker.FontSelectorDialog
+import sp.it.pl.gui.objects.window.NodeShow.RIGHT_CENTER
 import sp.it.pl.main.APP
 
 /** Text field for [Font] with a picker. */
@@ -12,7 +13,7 @@ class FontTextField: ValueTextField<Font>({ APP.converter.general.toS(it) }) {
    }
 
    override fun onDialogAction() {
-      FontSelectorDialog(value, { value = it }).showInCenterOf(this)
+      FontSelectorDialog(value, { value = it }).popup.show(RIGHT_CENTER(this))
    }
 
    companion object {