+++ src/player/main/sp/it/pl/gui/itemnode/textfield/MoodItemNode.kt
@@ -1,19 +1,18 @@
 package sp.it.pl.gui.itemnode.textfield
 
-import javafx.scene.Node
 import sp.it.pl.audio.tagging.Metadata
 import sp.it.pl.gui.objects.autocomplete.AutoCompletion.Companion.autoComplete
 import sp.it.pl.gui.objects.picker.MoodPicker
-import sp.it.pl.gui.objects.popover.NodePos
-import sp.it.pl.gui.objects.popover.PopOver
+import sp.it.pl.gui.objects.window.NodeShow
+import sp.it.pl.gui.objects.window.popup.PopWindow
 import sp.it.pl.main.APP
 import sp.it.util.access.v
 
-/** Text field for audio mood tagging values with a picker and autocompletion. */
+/** Text field for audio mood tagging values with a picker and auto-completion. */
 class MoodItemNode: ValueTextField<String>({ APP.converter.general.toS(it) }) {
 
    /** The position for the picker to show on. */
-   val pickerPosition = v(NodePos.RIGHT_CENTER)
+   val pickerPosition = v(NodeShow.RIGHT_CENTER)
 
    init {
       styleClass += STYLECLASS
@@ -24,14 +23,9 @@ class MoodItemNode: ValueTextField<String>({ APP.converter.general.toS(it) }) {
    }
 
    override fun onDialogAction() {
-      val p = PopOver<Node>().apply {
-         detachable.set(false)
-         arrowSize.value = 0.0
-         arrowIndent.value = 0.0
-         cornerRadius.value = 0.0
-         isAutoHide = true
-         isAutoFix = true
-         contentNode.value = MoodPicker().apply {
+      PopWindow().apply {
+         isAutohide.value = true
+         content.value = MoodPicker().apply {
             root.setPrefSize(800.0, 600.0)
             onCancel = { hide() }
             onSelect = {
@@ -40,8 +34,9 @@ class MoodItemNode: ValueTextField<String>({ APP.converter.general.toS(it) }) {
             }
             buildContent()
          }.root
+
+         show(pickerPosition.value.invoke(this@MoodItemNode))
       }
-      p.show(this, pickerPosition.value)
    }
 
    companion object {