+++ src/player/main/sp/it/pl/gui/nodeinfo/ItemInfo.kt
@@ -15,6 +15,8 @@ import sp.it.util.async.runIO
 import sp.it.util.identityHashCode
 import sp.it.util.ui.fxml.ConventionFxmlLoader
 import sp.it.util.ui.layFullArea
+import sp.it.util.ui.prefSize
+import sp.it.util.ui.x
 
 /** Basic display for song information. */
 class ItemInfo @JvmOverloads constructor(showCover: Boolean = true): HBox(), SongReader {
@@ -38,6 +40,7 @@ class ItemInfo @JvmOverloads constructor(showCover: Boolean = true): HBox(), Son
       if (showCover) {
          thumb = Thumbnail()
          thumb.borderVisible = true
+         thumb.pane.prefSize = 200 x 200
          coverContainer!!.layFullArea += thumb.pane
       } else {
          thumb = null