+++ src/player/main/sp/it/pl/layout/widget/controller/io/Input.kt
@@ -12,7 +12,6 @@ import java.lang.reflect.ParameterizedType
 import java.lang.reflect.Type
 import java.util.HashMap
 import java.util.UUID
-import kotlin.streams.toList
 
 open class Input<T>: Put<T?> {