+++ src/player/main/sp/it/pl/plugin/notif/Notification.kt
@@ -7,8 +7,7 @@ import javafx.scene.input.MouseEvent.MOUSE_CLICKED
 import javafx.scene.input.MouseEvent.MOUSE_ENTERED
 import javafx.scene.input.MouseEvent.MOUSE_EXITED
 import javafx.util.Duration
-import sp.it.pl.gui.objects.popover.PopOver
-import sp.it.pl.gui.objects.popover.ScreenPos
+import sp.it.pl.gui.objects.window.popup.PopWindow
 import sp.it.util.async.executor.FxTimer.Companion.fxTimer
 import sp.it.util.collections.setToOne
 import sp.it.util.functional.invoke
@@ -18,7 +17,7 @@ import sp.it.util.ui.stackPane
 import sp.it.util.units.seconds
 
 /** Notification popover. */
-class Notification: PopOver<Node>() {
+class Notification: PopWindow() {
    private val closer = fxTimer(5.seconds, 1, ::hide)
    private val root = stackPane()
 
@@ -36,30 +35,21 @@ class Notification: PopOver<Node>() {
       }
 
    init {
-      detached.value = false
-      detachable.value = false
       userResizable.value = false
-      isHideOnEscape = false
-      arrowSize.value = 0.0
-      arrowIndent.value = 0.0
-      cornerRadius.value = 0.0
-      isAutoFix = false
-      isAutoHide = false
-      skinn.setTitleAsOnlyHeaderContent(true)
+      userMovable.value = false
+      isEscapeHide.value = false
+      isAutohide.value = false
+      headerIconsVisible.value = false
+      focusOnShow.value = false
       styleClass += "notification"
-
-      contentNode.value = root.apply {
+      onShown += { closer.start() }
+      content.value = root.apply {
          setMinSize(150.0, 70.0)
          onEventDown(MOUSE_CLICKED, PRIMARY) { lClickAction() }
          onEventDown(MOUSE_CLICKED, SECONDARY) { rClickAction() }
          onEventUp(MOUSE_ENTERED) { closer.pause() }
          onEventUp(MOUSE_EXITED) { closer.unpause() }
       }
-   }
-
-   override fun show(pos: ScreenPos) {
-      super.show(pos)
-      closer.start()
    }
 
    fun setContent(content: Node, titleText: String) {